Fix worker bulk implementation & test worker action
