+++ .github/workflows/test_worker.yml
@@ -1,0 +1,35 @@
+on:
+  workflow_call:
+
+jobs:
+  test:
+    name: Test ENState Worker 🚀
+    runs-on: ubuntu-latest
+    env:
+      SCCACHE_GHA_ENABLED: "true"
+      RUSTC_WRAPPER: "sccache"
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+
+      - run: |
+          rustup set auto-self-update disable
+          rustup toolchain install stable --profile minimal
+
+      - name: Run sccache-cache
+        uses: mozilla-actions/sccache-action@v0.0.3
+        with:
+          version: "v0.7.4"
+
+      - uses: oven-sh/setup-bun@v1
+
+      - run: bun install
+        working-directory: test
+
+      - name: Test
+        run: bun test tests/worker.spec.ts
+        working-directory: test
+        env:
+          RPC_URL: https://rpc.ankr.com/eth
+          OPENSEA_API_KEY: ${{ secrets.OPENSEA_API_KEY }}
